fix: rename isPostsAreQuoted into isPostsQuotedByUser and moved to profile.ts
fix: move calculateSkip to helper.ts
